requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
browser-use>=0.9.0
python-dotenv>=1.0.0
playwright>=1.40.0
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.18.0
PyJWT>=2.8.0
cryptography>=41.0.0
